Rename: rename Python interface package to `cupdlpx`
