+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["scikit-build-core>=0.9.0", "pybind11>=2.11.0", "numpy", "scipy"]
 build-backend = "scikit_build_core.build"
 
 [project]
-name = "pycupdlpx"
-version = "0.1.0"
+name = "cupdlpx"
+version = "0.1.1"
 description = "Python bindings for cuPDLPx (GPU-accelerated first-order LP solver)"
 readme = "README.md"
 license = { text = "Apache-2.0" }
@@ -16,7 +16,7 @@ dependencies = ["numpy>=1.21", "scipy>=1.8"]
 build-dir = "build/{wheel_tag}"
 cmake.minimum-version = "3.20"
 cmake.build-type = "Release"
-wheel.packages = ["python/pycupdlpx"]
+wheel.packages = ["python/cupdlpx"]
 
 [tool.scikit-build.cmake.define]
 CMAKE_CUDA_ARCHITECTURES = "75;80;86;89;90;90-virtual"
