+++ python/cupdlpx/_core.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ._pycupdlpx_core import solve_once, get_default_params
+from ._cupdlpx_core import solve_once, get_default_params
