+++ test/test_interface.py
@@ -1,8 +1,8 @@
 import numpy as np
 import scipy.sparse as sp
 
-from pycupdlpx.model import Model
-from pycupdlpx import PDLP
+from cupdlpx.model import Model
+from cupdlpx import PDLP
 
 # construct a simple LP for testing
 m, n = 3, 2
